Add a script to find files that can cause the GetPartialObject offset bug
The `mtp-galaxy-find-files.py` script browses the content of a Samsung
Galaxy device through MTP looking for files that may cause an error
when retrieved with the MTP GetPartialObject command. The file must
have a size that matches the condition (size % 512) == 500.

The script parses the output of the `mpt-files` command. Here is a
sample invocation:

    $ ./mtp-galaxy-find-files.py
    Device 0 (VID=04e8 and PID=6860) is a Samsung Galaxy models (MTP).
    ID: 52, filename: IMG_20180305_114918_5.jpg, size: 1496052 bytes
    ID: 99, filename: IMG_20180305_114926_5.jpg, size: 1548276 bytes
    ID: 972, filename: IMG_20180305_115203.jpg, size: 1555444 bytes
    ID: 1666, filename: IMG_20180312_151120_5.jpg, size: 2514932 bytes
    4 file(s) may fail with GetPartialObject (total: 1688 files)

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>