Add a bug reproducer for the MTP GetPartialObject issue on Samsung Galaxy devices
This program expects a MTP object ID as argument that corresponds to a
file whose size matches the condition (size % 512) == 500. You can
invoke the `mtp-galaxy-find-files` script to identify such files on a
Samsung Galaxy device.

Here is a sample invocation when the program fails to read the last
sectors of a file stored on a Samsung Galaxy device. Be patient
because the failure happens after a one minute timeout.

    $ ./mtp-galaxy-bug 972
    LIBMTP_Set_Debug: Setting debugging level to 2 (0x02) (on)
    LIBMTP LIBMTP_Detect_Raw_Devices[695]: Device 0 (VID=04e8 and PID=6860) is a Samsung Galaxy models (MTP).
    Object 972: size = 1555444 bytes
    GetPartialObject: offset = 1554944 bytes, count = 512 bytes
    == FAIL ==

Signed-off-by: Christophe Vu-Brugier <cvubrugier@fastmail.fm>